TEncCu: remove for-loops over QP, make them simple expressions
No sane encoder would ever do that sort of redundant analysis.  There is a lot
of cleanup that needs to trickle up from here.  iMinQP and isAddLowestQP can
probably be removed, and all the slice and SPS fields for QP RD can be removed,
all the way up to the config options that enabled it.