entropy: clarify the SBac's bit counting mode
The SBac class has always had the ability to be a bit counter without any other
external data structures. With this change, the SBac defaults to being a bit
counting SBac until it is given a Bitstream object to write into.  The class
no longer accepts a BitCounter object, since it would only add more overhead
to the bit counting.

TEncCu no longer needs o be told whether it is writing into a bit counting
SBac or not, it can ask the SBac to find out.

The BitCounting class is only used for SEI writing, and may disappear in order
to remove the vtable from the critical path of entropy coding.