frameencoder: remove second encodeCU() pass over CTUs when SAO is disabled
This is a performance optimization, it allows the encoder to generate the final
bitstream of each CTU as it is compressed and cache hot.

When SAO is enabled, SAO analysis must be performed and coded at the start of
the CTU but SAO analysis currently requires surrounding CTUs to be encoded
making the second pass unavoidable.

Note that this commit changes the way non-WPP encodes are performed, for the
better. Now it always uses row 0's CI_CURR_BEST entropy coder instance to
communicate entropy state between all CTUs and between rows. This better models
how encodeSlice() works and makes RDO work better