encoder: pass cu and cudata by const reference when they are not modified
This is the beginning round of a sweeping refactor to enforce some const
discipline. The goal is to always pass objects as const references when a
NULL pointer is never allowed and when the object is unharmed. un-const refs
should be used when NULL is not allowed but the object is modified by the
function.

As we add more and more parallelism, we need to enlist the compiler's help in
figuring out which functions are safe to call without side-effects