cu: fix sub-cu pel positioning
The cuData.encodeIdx already included the sub-part offset during splits (there
is one cuData instance per sub-part, not per depth). So it was wrong to pass
the partUnitIdx to the next recursion depth.

It was only being used at the next depth to pass to initSubCU() which used it
to calculate the relative pixel offsets from the parent CU.  But in the new code
the parent CU is now the parent CTU, and so the pel offsets must be calculated
relative from the CTU pel positions. This is what the g_zscanToPel? tables are
for.