analysis: allow --b-intra to work with RD levels < 5 (no behavior change)
This changes no existing presets. This flag used to default to be on, but it
was ignored at RD levels 0..4. Now it defaults to off, but is enabled by all
the presets that use RD 5 and 6. This commit re-orders the tools log line to
place similar options near each other