analysis: try lossless encode of only the best mode for each CU
This involved some subtle changes in behavior. TComDataCU::initCTU and initSubCU
now initialize m_cuTransquantBypass to the global --lossless flag, so in the
first pass of analysis it will always be all non-lossless (typical) or all
lossless (when --lossless is enabled).

Before this change, when "--cu-lossless and !--lossless" the encoder would try
both lossless and non-losseless coding of every option evaluated for RDO.
This roughly doubled the work performed by the encoder, and because the same
TComDataCU instance was being used for both, it was fragile

After this change, the encoder will only try lossless once per CU using the
best mode found during non-lossless RDO, using a seperate TComDataCU instance.