predict: enforce calling conventions, fix wrong side-effects
use references and consts where possible, order arguments to follow the
convention of memcpy (dest, src)

This exposed a bug in addWeightBi() and addWeightUni(), they were modifying the
PU size variables directly instead of making chroma versions. This explains why
it seemed to best necessary at times to make seemingly redundant calls to
prepMotionCompensation.

as a side-effect, this commit also removes the 1k 'avg' buffer that bidir
allocated on the stack and instead uses the existing tmpPredYuv