encoder: give more warnings when features are automatically disabled
and add comments describing why the combinations are prevented.  Some of them
are simply impossible, the option would have no affect and so it is best not to
pretend it is enabled. Some will not be useful (have a negative impact on
performance with no compression improvement). And others are just currently
broken and not typically used.

--HG--
branch : stable