analysis: fix merge logic for RD levels 0..4
At the end of this function, md.bestMode points to the final 'best' merge mode
for the CU but the end of the function was filling in CU data for bestPred,
which could have been the wrong one.  This left MV info in the CU uninitialized
and led to non-deterministic behavior.

--HG--
branch : stable