frameencoder: use atomics to determine full completion of wpp and filter work
This is a more complete solution for the problem described in 7b5d44d7831e.
Since that commit we've seen one encode failure with the same missing four
bytes, demonstrating the race hazard was still possible (though particularly
hard to reproduce).

in this solution, we increment a counter as each row is first compressed and
again when it is filtered and measured for metrics.  In this way we cannot
signal completion until all rows are completely finished, and it should be
impossible for the signal to occur twice.

--HG--
branch : stable