threading: intrdoduce poke() method for ThreadSafeInteger
Sometimes we need to waken all threads that are blocked on a TSI,
but do not want to change its value (we want the blocked threads to
re-check other state variables). We were using set(get()) for this
but that is a race hazard.  We've seen one deadlock in the vicinity
of this code and so are removing this hazard even though we are not
certain this was the root cause of the deadlock.

--HG--
branch : stable