hdr: revert calcHDRStats
This was developed with the goal of calculating the metadata extensions specified
by the CEA 861.3 specification(from the source YUV). However, we realised that
the specification is meant to work on RGB data, and transforming YUV (back) to RGB
domain will yield very different results. So, as of now, the user is required to
give the maxcll/maxfall arguments (generated as per recommendations) and x265
will generate the appropriate SEI messages. (Shout out to ColorFront team for
their helpful clarifications).

--HG--
branch : stable